Fix content-sizing for images.
Images normally expand to fill (or fit) their limits, so there is an inherent conflict for content-sized dimensions which say to ignore the limits.

There isn't a clean way to work this into the `LayoutParameters` so it's just a few conditionals in the `Image` itself instead, to override the limits.