Add a Panel for overlapping/overlay views.
Views to be "stacked" according to Z order which allows for overlays. We're using this right away for the tab icons; while tabs _could_ also be a `Frame`, it's inconvenient as the border and background aren't really separate, and is just a single sprite anyway.

This way is much simpler for tabs and will probably be used for other widgets later.