Add grid view layout and grid-based item chooser.
We'll use this to allow adding items to the manifest of a single node. May implement this as a checkbox UI but for now it's just a grid.

For now, adds a bunch of sample items. Eventually this must be changed to look at connected storage and store all current/previously seen items - since listing every single item in the game would be confusing.