Implement focus search for view menu and item grid.
Fixes several bugs and incomplete implementations, mostly to do with position propagation (there has to be a cleaner way to do this?) and a few other logical errors. Also overrides the required `IClickableMenu` method to actually implement it for the menu.

Makes `IsFocusable` a settable property so any view can become focusable. Original idea was to have "buttons" and other implicitly focusable components, but not sure there's much value in adding more layers.