Make Lane focus search less sucky.
Rewrites the implementation to be simpler (sort of) and more reliable (definitely). Has explicit rules for constraining the search to not leave an existing focus area.

Also updates the Grid to allow along-orientation movement to pull back to the previous row/column in order to keep going, i.e. as opposed to getting "stuck" in the middle of the bottom row or the right column.

The primary test case for this is making the arrows at the top of the node view focusable and making sure focus moves smoothly between the different areas of the UI.