Support horizontal tabs.
Since the game does not have a blank horizontal tab sprite, we can get this by rotating the vertical sprite; the actual pixels are identical.

XNA makes this a bit of a kludge because of the way it treats "origin", but this can be encapsulated in the `NineSlice` and not leaked anywhere else.