Start logging focus searches and naming all views.
For this type of recursive logic, reading a clear log is often much easier than stepping through execution.

Includes fix for one of the main focus-search bugs, where Panel would not finish the search unless the point was within the panel's bounds. This is incorrect for a _search_, we only want to know if the bounds are in the correct direction, but should still _prioritize_ the child that already contains the cursor (if there is one).