Refactor and perf-tweak click handling and view-path construction.
We don't need to require a search of all children in all cases; some view types like the Grid know exactly what _could_ be at a given position and can return a result in O(1) time.

Even in other cases like Lanes it might be possible to optimize with a binary search if the child positions are known to be sorted along an axis - but child lists aren't large enough in those views to make a big difference right now.