Add tab selection for current node.
Refactors tab views into a more general "tab group" view which can be bound (sort of) to an enum or other key.

Coordinating property setters with lazy-initialization is proving to be a pain, but no obvious solution has presented itself (yet).

This makes either the rule table or item grid selectable within the current node.