Fix size and syncing of scrollbar.
Updates the `Lane` to support views that stretch to "used" width/height, i.e. on the perpendicular axis. Defining some views with fixed height and other views to stretch is a common use case. Scrollbar will use this so it tracks the height of the container.

Improves synchronization between the scrollbar and its container so that it's a mostly hands-off process and changes to the container are automatically reflected. This also makes the visibility of the scrollbar consistent with whether or not there is anything to scroll.