Rewrite `ScrollIntoView` so that it actually works.
Previous implementations half-worked but were fundamentally broken because they only looked at the specific focused element.

Instead, this implementation tries to bring the entire "row" or "column" into view, which in practice means traversing back up the tree, accumulating union bounds until the descendant would be too large to fit, then scrolling to cover those bounds, plus a "peeking" factor to help with margins and making adjacent content discoverable.

Unsure if this covers every possible edge case, but for the simple cases being used and anticipated, it's effective. The implementation also returns the distance scrolled, which is used to reposition the cursor so that it stays on the scrolled element. (Note however that this might be moved outside the scrolling API and into an event so that it also covers wheel scrolls)