Maintain cursor focus while scrolling.
Only applies when controller is used for navigation. If the previously-focused view is still in view, the cursor should move to the center of that view; otherwise, we guess where the next best point is by doing a focus search in the scroll direction.

This works on its own but also helps with `ScrollIntoView` to make sure the cursor doesn't get into a weird spot. Behavior is likely very different from some in-game menu behavior, because it has to work with arbitrary positions, i.e. we aren't currently forcing wheel-based scrolling to use an integral number of "rows" because the scroll container lacks that info. However, it should still be relatively intuitive - focus stays on the focused element if possible, otherwise snaps to the nearest visible position that is focusable.