Add hover animations for controls.
This is implemented via a semi-generic animator, and a mostly-general mouseover/mouseleave system (called pointer here to reflect controller input) comparable to mainstream frameworks.

There's a curious issue where the enter/leave events don't trigger for scrollbar arrows specifically when the previous/subsequent position is inside the main frame; this still has to be diagnosed but everything else seems to work correctly.