Fix scrollbar container width.
The reason for the hover state getting "stuck" was overlap due to the view exceeding its container bounds; the mouse handler was getting confused, because cursor position left the parent (scroll bar) _before_ leaving the child (arrow button) and so the final mouse-out event had no chance to propagate down.

There's probably a more complex solution that would accommodate such overlap, but it's simpler to just fix the dimensions so the scroll arrows don't have negative positions, and more consistent with how existing game UI already looks with the scrollbar farther to the right.