Add label ellipsis and set length limit for ruleset name.
The length limit should fit comfortably for "typical" names, and the text box is about as large as it can get without starting to look strange.

It's still possible to exceed this with names like "WWWW WWWW WWWW WWWW" which is where the ellipsis comes in, which helps indicate that there is actually more text (out of view), better than having the word mysteriously disappear.

While a full-fledged implementation could potentially scroll, or allow multiple lines, it's considerably more complex to implement those for scenarios which should be very rare.