Add a "shipment report" view.
Designed to be used either as standalone, or hosted within the Logistics View where there might also be navigation by date. The standalone version is for testing, dry-runs, etc.

Not finished yet; missing clear import/export indicators (arrows?), some focus-search is buggy when switching between the two columns, some non-localized text, flavored items draw generic images, etc. This is a rough draft.

Adds a "percent" length type, but this doesn't work entirely as expected, since it's a percent of the *remaining* length during layout, meaning a series of 50% elements keep getting smaller and smaller. However, it's enough to solve the immediate problem of having a 2-pane split layout as the 2nd one can be set to Stretch.