Fix scrollbar alignment.
`Middle` was being used to align track with arrows, but also meant that if the scroll bar's "container" was given a large width (i.e. to keep the scrollable container centered when using an opposite sidebar), then the scrollbar itself would be pushed to the right.

Pixel offsets seem OK for this because the arrow dimensions are already hardcoded.