Absorb `ScrollableFrameView` into UI library.
This is used in a few projects now and the implementations are similar enough to be able to define a single class to handle them all. It's a common enough scenario that it justifies a shared implementation.

Since the code already depends on some of SDV's types, the actual sprites have finally been added directly to the project. There's a risk of scope creep, but should be OK as long as it is constrained to "sprites necessary for basic UI". Externalizing every single sprite for the frame view is just too onerous on the caller and generally pointless right now.