Add deferred stretching to `Panel` and incorporate minimum layout sizes.
This is probably not a perfect solution to handle `MinWidth` and `MinHeight` but it's better than nothing. This way it's possible to combine stretched and unstretched (fixed/content sized) children in a panel, and have the stretched children match their lengths to other children, while still being able to enforce minimum size on any of them, especially e.g. an `Image` that looks bad at tiny size.