Replace accidental "iterative" `ViewMenu` measurement.
The menu's `width` and `height` are calculated _from_ the view's size, so subsequently using those dimensions as limits can lead to rounding/text-wrapping errors and other strange inaccuracies that progressively shrink the menu size.

Instead, do the same calculation that's done in the constructor; always use the full viewport size as available size and always re-center if necessary. The cost of this should be relatively inexpensive, it doesn't affect dirty checking and should actually reduce extra measurements in the cases where these tiny errors were accumulating.