Ensure fixed dimensions are taken into account for Panel child layout.
Deferring stretched children works, but if the panel itself is not content-sized in that dimension, then the children may be given too small a limit size.

Simplest solution is to re-resolve using the `LayoutParams`.

Also adds a `Panel` factory helper for a common use case, adding multiple panels to "dock" different children of a parent `Panel`.