Use "actual content bounds" instead of "actual outer bounds" for focus searches.
When looking for the center of some `ViewChild`, we do not really want to include the margins, as they are not perceived as part of the view.

This still takes into account negative margins (which shift the content rather than leaving empty space), but will exclude the empty space in positive margins.

For the moment, padding is not taken into account - it might need to be later.