Add overlay system and drop-down list.
The drop-down is based on the overlay. Overlays can position themselves relative to the screen, or to any other view; there is a stack of overlays that are all drawn, and only the one on top receives input events.

Can close overlays (and dropdowns) by clicking outside or by pressing B button (keyboard ESC/E), it won't close the menu if an overlay is open.

For gamepad, closing an overlay will trigger an attempt to return focus to the parent that positioned it, which does the right thing at least for dropdowns. May need tweaking for other scenarios.