Remove source-size dependency in Image dirty check.
No longer remember why this was added in the first place, but it breaks the default behavior of the checkbox which switches between two equally-sized, content-sized sprites.

It doesn't seem logical to have the condition anyway; if the sprite changes size and the view _is_ content-sized, then the sprite changes is by definition a layout change. The previous logic simply seems wrong.