Apply non-content-deferred swap-orientation fix to `Lane`.
This was previously applied to `Panel` in order to get stretched children to actually stretch to the full length when the lane length (in that dimension - the swap dimension) was not `Content`. Stretched children should only "shrink" to other children when the entire lane wants to fit the content.

The logic is slightly different for lanes due to all the orientation-specific code.