Try to maintain gamepad cursor position during major layout changes.
If a significant change happens, particularly one that causes scrolling or even shifts from a non-scrollable to scrollable layout (or vice versa), then (a) the previous scroll offset can become invalid and (b) the cursor position will no longer make sense.

This change makes refocusing an implicit part of the menu's measure loop (if the layout was dirty) in addition to the explicit refocus on scroll wheel, which should resolve most of these issues.