Add a floating-element system.
This is early, and possibly buggy, but is designed to handle a use case that's otherwise rather painful and requires weird hacks like negative margins.

Obvious use cases are scrollbars and sidebars. These are meant to have their own independent layout, not really participate in the parent layout; i.e. we want to have the "main content" view perfectly centered and be able to add a header, footer, side tabs and scroll bar without having to do a lot of silly things with margins and pixel widths to make them all cancel each other out.

Nothing is converted to use this yet, but it's been tested to work with mouse hover/click/drag events. Focus search uses a naive implementation and may have issues.