Add a `ScrollableView` based on floating layout.
This is designed to eventually replace the implementation of `ScrollableFrameView` and handles a wider variety of use cases as it doesn't impose any assumptions about where the container is hosted. Floating layout ensures that the scrollbar is (or should be) always in the right place; always perfectly vertically scaled with the actual size of the content view and however far right it needs to be (can be adjusted for different sized borders). No multilevel nested layouts with margin hacks required.