Use outer size to measure floating elements.
Content size seemed like a good idea but in practice just seems to make everything harder to align. For the scroll bar it makes no difference at all, and for headers/footers, content size overlapped with the border.

Using outer size, there may be a need to copy margins (etc.) from the parent element to the floating element, but that is generally going to be one line of code, compared to a lot of fiddling trying to "undo" an unwanted margin.