Add separate left- and right-click events.
We are using `Click` everywhere and _never_ remembering to check whether it is the primary button; clearly, this does not scale well.

To avoid breaking old code, `Click` still behaves the same as before, but this change adds `LeftClick` and `RightClick` events with the expectation that most handlers use `LeftClick` specifically. All internal widgets have been converted to use `LeftClick`.