Add entry point for SMAPI/logging init.
Previously mods would set `Logger.Monitor` directly for just logging. They can still technically do that, but from now on they are going to be required to call `UI.Initialize` which also sets up logging (so the former call can be removed).

With the ongoing keybinding work, having access to the SMAPI input helper (for input suppression) is too important to let go of; the vanilla APIs are just way too clunky.