Rework relationship between reported buttons and actions.
Stardew's reporting of buttons/keys to the various `IClickableMenu` functions is a mess. In addition, an older change to the `ClickEventArgs` reducing the logic to just `IsToolUseButton()` and so on, was actually wrong, because the world-interaction button groupings are different from the UI-interaction button groupings.

There is no coherent spec or set of rules for this anywhere, so this change adds its own through `ButtonResolver`. Whenever possible, menu handlers will try to locate the "real" button that triggered a menu action based on what is actually down (held) at the time, and instead of checking specific `SButtons`, our own UI methods will check the `ButtonAction` associated with them.

Perhaps most importantly, this allows correct use of input suppression via the SMAPI input helper, which is necessary to prevent "leaking" inputs between overlays or between the menu and game world.