Add an input capturing system for overlays.
While the need for this is probably rare, it is absolutely essential for the keybind overlay since it needs to steal all normal navigation/confirmation/cancellation input from the underlying menu.

When an overlay is in capturing mode, none of the normal UI input handling runs and the overlay is entirely responsible for letting go of the capture at some point to return control.