Implement the keybind editing overlay.
Currently only supports one type of keybind, but it is the most complex type (multiple keybinds). The others are narrower cases and should be easier to implement.

Tested a decent amount, with both keyboard and gamepad bindings. Cancelling out of the capture state can be done by either finishing the keybind or hitting Gamepad B / Keyboard Esc.