Improve default and return focus for menus and overlays.
Views can now specify their own default focus, which has a baseline implementation identical to previous behavior (first focusable descendant) but can now be customized, which is especially useful for e.g. jumping straight to the selected option of a dropdown list, which is now also implemented.

In addition, `ViewMenu` now saves the "activation path" of an overlay - that is, the exact path to the UI element where the cursor was located prior to the overlay being shown - and returns the cursor to the same _element_ after the overlay is closed. This means no having to awkwardly reposition the cursor after a drop-down, popup, etc.