Add an update/tick method for views.
While this should be (and is, in doc comments) heavily discouraged, in some cases it really is necessary, e.g. a custom view that must handle its own input as part of an overlay, or sync with some other game state.

Removes spammy log line from ViewMenu, which was left in accidentally.