Add dirty checking for content alignment.
Changing content alignment won't change the dimensions of the layout view, which is probably why it was excluded from dirty checks originally; however, it very obviously does change the position of those views, and computing the position is still part of the measurement/layout. Therefore we need to consider these dirty, and changing them needs to trigger relayout.