Round out input prompt sprites and add a general `SpriteMap`.
Includes (in the binary changes) some individual sprites that aren't useful for key bindings but are going to be useful for positioning prompts - stick motion, unpressed d-pad, and mouse buttons.

Also adds arrows which are very helpful for suggesting a drag motion (vanilla cursor arrows look too much like navigation), and a `SpriteMap` as a convenient way to build the model for those arrows and probably many other texture atlases in the future.