Early work on positioning overlay.
Gets a lot of the basic functionality working, including the button prompts and actual alignments, both nudge and snap. It is not completely done - for example, mouse dragging is not implemented at all, and the mouse cursor really should be hidden entirely for this view when using gamepad controls (no clean mechanism to do that yet, as capturing mode would hide the mouse cursor as well).

Also needs a whole lot of refactoring to actually follow the `ControlScheme` and refactor some other modular pieces like button-repeat and throttling.

But committing this now in case it gets badly broken during subsequent updates.