Refactor: Extract keybind handler from `PositioningOverlay`.
This long-overdue class is the localized equivalent of SMAPI's keybind system + all the hacked up mod code using it. It's a central point (intended for _local_ use, as in the PositioningOverlay) to register keybindings for some specific UI and have it automatically handle the press detection, action mapping _and_ repeat/suppression behavior, the latter of which SMAPI does not provide and is otherwise completely hardcoded in `Game1`.

Part of what will likely be a long series of refactorings for this UI.