Extract "nine-grid" operations into utility class.
These are all currently in service to the positioning overlay but are otherwise independent of it; the logic around layout, neighbors, snapping, etc. can be and soon will be applied to the actual HUD as well as the config UI.

This eliminates a very large chunk of code from `PositioningOverlay` and makes it almost readable.