Use `ActionState` for keybind display in positioning overlay.
This _mostly_ resolves to the registered key bindings, although it is difficult to do correctly for the d-pad/stick nudge movements because those each use a single sprite to represent all the binds.

Keyboard nudge and grid snap use the action state correctly and the overlay has been updated to use `KeybindView` instead of the hacky substitute.