Implement drag-and-drop in positioning overlay.
Getting to the intended look and feel requires an extra overlay on top of the real view, which requires being able to dispatch pointer events to the views underneath. To make this work, the change includes a `PointerEventsEnabled` that is analogous to CSS `pointer-events`, essentially blocking any view and its descendants from receiving clicks/drags/etc. **and** ensuring that they reach any views underneath (overlapping but with lower z-index).

Dragging in particular is still a little wonky (see code comments) but works for this use case.