Improve support for submenus/title menus.
Provides a way to run custom actions on close (e.g. to open up a different menu, which for title submenus has to be custom and cannot use `nextClickableMenu`), a workaround for possible `Dispose` loop, and a convenience method to open/close a created menu which works correctly from the title screen.