Fix buggy scrolling when focus target has position offset.
`ViewMenu` was coded to always focus to the end of the `FocusSearchResult.Path`, which is the path to the _parent_ of the actual focus target; thus the actual focus search "operation" only ended correctly if the leaf had a zero position within its parent along the scroll axis.

Pulling on this thread caused the whole `FocusSearchResult` to unravel, as for reasons no longer apparent its offset operations had been coded to offset both the target _and_ the first path element, when the correct logic should be the target _or_ the first path element (the former for empty paths, obviously, and the latter for non-empty).

It seems all prior code had been written, coincidentally, with focus targets in some `Frame` or `Panel` such that this never got noticed.

This fixes the bad offset logic in `FocusSearchResult`, and updates `ViewMenu` to include the final target when resolving the global position. It also fixes a minor nuisance in the `Refocus` implementation that manifested as "instability" in the focus, caused by not adjusting the search origin for the _new_ position of the previously-focused (and now out-of-bounds) element before doing a focus search.