Use discrete binding modifier tokens.
For asset bindings, instead of on-the-fly checking for the `@` prefix, the lexer/parser handle an explicit binding modifier between the binding start token and the binding attribute value, which determines default (input) vs. asset vs. explicit directions.

Output and 2-way bindings are not implemented on the binding side yet, this only opens up the grammar for their use.

#1