Refactor: Change `IViewNode` to have a list of views.
Structural attributes (may) need to emit multiple views. It doesn't seem sufficient to have this reflect through the `ChildNodes`, because even though the nodes themselves can be resolved meaningfully, the singular `View` at the node level would be unclear, especially in terms of layout. There is really no such thing as a "pass-through view" and cannot be.

#2