Add/implement `*context` attribute.
To sidestep a lot of the ambiguity around attribute priority, this attribute only affects _children_ of the annotated node. If it ever becomes completely necessary to rescope a "current" node then this could be done using some sort of pass-through tag that behaves like a `WrapperView`.

Also, because it is getting confusing to pass `object` types around everywhere, most context properties are switched to use `BindingContext` now, which will soon be extended to include the ancestry.