Implement depth-based context walking.
This is implemented in the `ValueSourceFactory` which everything else uses for bindings, so it should be automatically applicable to anything that uses a context binding.

Test notes a quirk - that the walked bindings don't resolve unless the "current" binding exists. For now this is WAI, if not totally ideal. Probably the only other practical alternative would be enforce non-null `BindingContext` but allow nullable properties inside them.

Changed `Font`/`SpriteFont` assertions in tests to use a different property (`int MaxLines`) because static fonts are all `null` in tests, making the assertions meaningless.

#3