Adds name-based context redirects.
Allows walking up to a parent context with any given (short-form) type name.

Requires a tiny stateful hack in the lexer to avoid ambiguity between literals that should ignore periods as separators (namely, `@asset` bindings) vs. those that sometimes now need to separate on them (context bindings). Although the parser could be modified to re-join dot-separated tokens in `@asset` and similar bindings, it would be an unnecessary slowdown as they are already part of a single span in the input stream.

#3