Update test mod to use file-based asset.
With the Harmony patch in place, the entire `AssetRequested` handler per mod can be written in a few lines, so there doesn't seem to be a pressing need to do extra work (and complexity) in the API for a minuscule ergonomic improvement.

Content Patcher should be very similar (though untested as of now), except of course file-by-file rather than a wildcard handler as in this SMAPI example. The patch has been updated to allow `object` as the target type, since `LoadFromModFile` requires a generic param and API consumers don't have the `Document` type.

#6