Implement included views.
Creates a new node type, instead of a new view type, because that way the node hierarchy doesn't get "split" across an arbitrary view boundary and it is easier to inherit context, reference a parent context, etc.

Uses a `name` attribute so that it is possible to change _which_ view is included via the binding syntax (as compared to taking a `Document` and providing something like `<include content={{@Views/View}} />`.

Uses the `IAssetCache` abstraction for easier unit testing and to handle asset invalidation (hot reload).

Tests should cover the majority of happy paths; it's getting impossible to handle every possible combination of elements and attributes in the test suite, but the most difficult aspects are working with repeaters, changing contexts and asset invalidation, which are all tested.

#4