Add grammar for event bindings.
Includes both the raw types on the grammar side and parsed types on the DOM side, as well as various reworks on the lexer and parser to handle the new and very different "sub-grammar" of an event which is more complex than regular attributes.

This is only the syntax, and does not include a runtime implementation for invoking the events.

#5