Extend event argument syntax to include event object properties.
This will, once bindings are implemented, allow passing properties of the `EventArgs` subclass like `ClickEventArgs` via a prefixed name such as `$Button` or `$Position`. The idea is for this to be bindable as long as either the type is identical (`SButton` and `Vector2` can cross the API) or has a supported type conversion (e.g. `Direction` enum to a `string` argument in the handler).

#5