Implement event bindings, first take.
This involves new descriptor and binding types and some significant rework on the existing types.

The number of variables interacting with each other to make this work is enormous (two different method signatures, types and conversions for each argument, separate contexts for the view, handler and arguments that are all allowed to change, etc.), so although we have exactly one test passing, there may be many other undiscovered problems.

#5