Support drop-down and tiny-digits views in markup.
Tiny digits was a matter of adding default sprites, which is trivial to do - just copied from another project.

Drop-down list is difficult because it uses a generic type but the view type cannot be chosen dynamically based on the binding data, as that's the opposite of how the binding tree is actually constructed. A clumsy but working hack for now is to just use `object` for the item type, meaning the model would have to do its own casts.

There might be ways to improve this later; possibly implementing more conversions between list types, and/or defining an internal "any" type that claims to be convertible to anything and does all its checks at runtime. For now, this is just getting the widget up and running.