Add asset registry.
Handles both view registration (as a convenience) and sprite registration (as more or less a necessity).

Mods can now register their paths through the API, which uses a bit of reflection to pull the mod's own content helper/events in order to use its own directory.

Adds the proposed JSON schema and data/texture (json/png) pairing system, working with both referenced/vanilla assets and mod assets in the same directory. Although it might be a bit unintuitive, for now it is the texture and not the data that receives the expected asset name like `Mods/MyMod/Sprites/SpriteSheetName`, because this way it is possible to reference that texture like any other game texture such as `LooseSprites/Cursors`. Whereas there should really never be any need to reference to the JSON asset directly (vs. the sprites in it).

#7