Convert `DocumentView` into wrapper.
Replaces the `Document` reference in the constructor with an `IValueSource` that can monitor changes to the asset, and recreate the node and/or root view as needed.

Essentially behaves a lot like the `WrapperView`, but since the view can change after creation in hot-reload scenarios, it needs a bespoke implementation.

#8