Implement automatic hot reload.
Reloading views is supported by the `DocumentView`, but to give the desired "experience" we need to provide a way to invalidate the assets. This does so by actually monitoring the file system to provide instant updates as soon as the file is saved.

In order to work, the assets have to actually be managed by the mod, i.e. in the `AssetRegistry` - if mods provide their own asset request handlers then they have to implement their own invalidation.

Also works for sprites and sprite sheets now (required some changes to how they are cached).

The feature is opt-in via the API (`EnableHotReload`), and to get true "instant" reload without having to alt-tab to the window, users will need to turn off the setting that pauses the game while the window is inactive.

#8