Change most double braces to single.
There's really no point in requiring double, it's of no benefit to the lexer/parser; it's just similar to how Content Patcher mods work, but that is because those double braces are going _inside_ quoted attributes whereas we make quotes the actual attribute boundaries to denote different attribute types.

Double braces are still allowed (and tested in a few places) because it may be a force of habit for some modders, but the recommendation is now single.