Fix: Remove expired entries from asset cache.
We were invalidating them (setting their `IsExpired` to `true`) but _not_ removing them, causing all sorts of thrashing as soon as any asset actually became invalid.

They need to be removed from the internal cache so that the cache wrapper actually creates a new, non-expired entry on the next request.