Add handling for `this` property to bind to self.
There will be cases, especially with repeaters, where we'll want to bind something to the actual context instead of one of its properties. A good example of this might be `ParsedItemData` to pass into an `Image` sprite.

Two-way/output bindings are obviously impossible to do this way but it will work for standard input bindings.