Fix leak/invalid cache in event bindings.
The factory method was written to accept the view context and handler context as separate arguments, but only the view context was actually being updated each time; the handler context was being (incorrectly) captured in the delegate, resulting in future creations of the menu reusing an old/expired content on which to run the event handler.

No design changes required, just fixing the bug so caching works as it was actually intended to in the first place.