Add backoff and logging for node updates.
When attempting to update a node throws an exception, instead of passing the exception and allowing it to happen again immediately on the next update, the framework will now log the error (non-fatally, so other UI might still succeed in updating) and use exponential backoff to wait for the next retry.

Hot reloading generally recreates the failed node, so the fix is actually immediate regardless of any pending backoff. While this is a positive, the safety is aimed less at hot reload and more at the potential of invalid bindings happening due to unexpected model state "in the wild", and the extended logging (showing exactly which node failed) should help significantly with identifying those issues.

This takes care of the 3rd case in #11. Still a bit more work needed for 2nd case.