Add backoff + error logging to `DocumentView`.
Will display the specific node that failed (if known) and make the exception non-fatal - although updates cannot propagate to the UI itself if the node cannot even be created, the menu can still be dismissed, and it will recover in a hot reload scenario if the problem is fixed.

Also remembered that we already had a `Logger` helper and there was no need to be passing `IMonitor` around everywhere or using the special patcher static field.

Resolves #11.