Map handler results to event `Handled`.
Used to prevent bubbling through returning a `bool` from an event handler; requires some bug fixes to `EventBinding` in order to actually accept a `bool` return type since the delegate cannot be cast to an `object` delegate.

Errors were initially hard to track down due to recent changes for #11 to suppress and log certain errors, so added a test implementation of `IMonitor` to ensure the output gets redirected to to the test result.

Fixes #19