Extract common base class for view decorators.
`DocumentView` and `WrapperView` are very similar in a lot of ways, but had some subtly different assumptions that led to two totally different implementations, and a third one was likely to emerge for the drop-down proxy.

An analysis of the behavior and usage of `WrapperView` also showed that the lazy-loading behavior was somewhere between useless and harmful; creating a view is not that expensive in the first place, and lazy loading might delay it by one frame, while vastly complicating the API and leading to inconsistent and erratic requirements for triggering the lazy load (some code would force load, other code would try to handle delayed loads, etc.).

The new base class, `DecoratorView`, is sort of complicated in how it handles properties and view assignment, but isn't meant to be used directly. Instead, `WrapperView` is renamed to `ComponentView` and creates the view in its constructor, so subclasses can write simple wrapper properties and not worry about the state, and `DocumentView` retains the same API but is now a much simpler subclass of `DecoratorView`.

Included: Cleans up a bunch of stray `using` directives and other formatting errors that somehow crept in from VS auto-include.

#9