Remove ctor params from all standard widgets.
Without the lazy-loading mess, this is now a lot more feasible; converting to properties is often just wrapping the same property on an inner view, or in some cases a small bit of extra work with a backing field.

This should make nearly everything customizable (and instantiable) from StarML - although turning `ISpriteMap` into an asset binding, which is the ideal use of these in StarML, is not yet supported.