Add `DynamicDropDownList` and switch framework.
Adds an `IAnyCast` type that induces forced explicit casting, so that the actual reflectable types on the `DynamicDropDownList` appear convertible to any binding.

The list will attempt to make changes immediately, like the `DropDownList<T>`, but ignore mismatches until the final end-of-frame update. This is done so that "good" updates generally happen within one frame as they are supposed to, and only "bad" updates have to wait a frame.

Slightly tweaks the order of updates in `ViewNode` such that _all_ input bindings are processed before _any_ output bindings. This way, no matter whether the `SelectedIndex` or `SelectedOption` gets visited first, the other is always written back to any output/two-way bindings in the same frame.

#9