Instrument various parts of the library/framework for tracing.
Coverage is not exhaustive, but already shows some very interesting (if not exactly surprising) results, e.g. `ReflectionViewBinder` and its related binding factories being _much_ more expensive than the descriptors themselves, and even some of the measure/draw calls being rather expensive the first time.

Refactors the main `Trace` helper into the Core library so that it can be used from the core library, and updates `TraceData` to reuse frames with the same name so as to add a lot less bloat to the trace files, which can already accumulate to several MB within just a few seconds given everything that is being traced.

Resolves #23