Add button-press event to views.
The semantics are unintentionally very similar to a `Click`, but the triggers are of necessity very different. This is meant to support a more general version of tabbing, paging and other such behaviors, although it will run for all buttons including those used for clicks and navigation.

Button press events can't be used to suppress default behavior (e.g. prevent navigation on a left or right d-pad press), at least not for now, though they should work as expected with bubbling.

Updates md docs, and cleans up a few doc comments that were worded poorly.