Update Core namespaces to match paths.
Throwing everything in `StardewUI` (root) is disorganized and a liability for documentation, among other things; it made sense when there were a dozen classes, not when there are a hundred.

This change unfortunately reveals an atrocious set of interdependencies between namespaces, but cleaning that up will have to wait for another day.